MBP 2015-12-02 Renames files, updates links accordingly, adds comments, cleans indenting. End of cpfrontend branch integration.
