+++ openSession.php
