--- testCheckLoginTest.php
@@ -1,8 +1,0 @@
-<?php
-
-include 'checkLogin.php';
-
-$test = checkLogin('andrewcaya', 'password3');
-$firstname = getFirstname('christianpineault');
-
-var_dump($test, $firstname);
