--- testMiddleLoginTest.php
@@ -1,9 +1,0 @@
-<?php 
-
-//Variables test que nous obtiendrons par le formulaire.
-$user = 'martinbeaudry';
-$password = 'password3';
-manageLogin();
-
-
-?>
